Create Compare_list_UMIs_AA_NT.py
I wrote this script to compare two sets of alignments. 
Because we sequenced varoius timepoints and made changes to some of our alignments, I needed to do a 'sanity check"
We generated amino acid alignments from NT alignments. I wanted to make sure that our AA alignments do not contain sequences that are NOT in the NT alignments, because that would be an indiation that I was working with the wrong set of files